% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{execute}
\alias{execute}
\title{Execute the Study}
\usage{
execute(
  databaseDetails,
  siteId = "site_name",
  uri,
  secret,
  outputFolder,
  createCohorts = F,
  createData = F,
  sampleSize = NULL,
  createControl = F,
  siteIds = "",
  leadSiteNextStep = F,
  runAnalysis = F,
  runSynthesize = F,
  verbosity = "INFO",
  cdmVersion = 5
)
}
\arguments{
\item{databaseDetails}{The connection details and OMOP CDM details. Created using \code{PatientLevelPrediction::createDatabaseDetails}.}

\item{siteId}{The name of your site (can be the university name, site with a number, etc.) tnis needs to be shared with the study administrator}

\item{uri}{The Universal Resource Identifier for the cloud}

\item{secret}{The password to authenticate as siteId on uri}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/). Do not use a folder on a network drive since this greatly impacts
performance.}

\item{createCohorts}{Create the cohortTable table with the target population and outcome cohorts?}

\item{createData}{Create the labelled data set (this is required before runAnalysis)}

\item{sampleSize}{The number of patients in the target cohort to sample (if NULL uses all patients)}

\item{createControl}{(for the lead site only) Run this code to create the study control}

\item{siteIds}{(for the lead site only) vector with the names of all the sites contributing to the study. Required when createControl = TRUE}

\item{leadSiteNextStep}{(for the lead site only) Run this when you are ready to move to the next step}

\item{runAnalysis}{Runs the initialization, derive and estimate.  This needs to be run multiple times - the study administrator will inform you when you run.  This step 
involves downloading the latest control json (with model specifications), fitting the model locally and then uploading the model (coefficients and ?) to 
the cloud for the study administrator to combine.  Please note: no patient level data are transferred.}

\item{runSynthesize}{Once the site estimates are returned, it is now possible to apply each model to the data to calculate
predictions.}

\item{verbosity}{Sets the level of the verbosity. If the log level is at or higher in priority than the logger threshold, a message will print. The levels are:
\itemize{
\item{DEBUG}{Highest verbosity showing all debug statements}
\item{TRACE}{Showing information about start and end of steps}
\item{INFO}{Show informative information (Default)}
\item{WARN}{Show warning messages}
\item{ERROR}{Show error messages}
\item{FATAL}{Be silent except for fatal errors}
}}

\item{cdmVersion}{The version of the common data model}
}
\description{
Execute the Study
}
\details{
This function executes the dGEMcovid Study.
}
\examples{
\dontrun{
connectionDetails <- createConnectionDetails(dbms = "postgresql",
                                             user = "joe",
                                             password = "secret",
                                             server = "myserver")

execute(databaseDetails,
        siteId = 'site_1',
        uri = 'get from administrator',
        secret = 'hidden',
        outputFolder = "c:/temp/study_results", 
        createCohorts = T,
        createData = T,
        sampleSize = 10000,
        createControl = F,
        siteIds = c('site_1', 'site_2'),
        leadSiteNextStep = F,
        runAnalysis = F,
        runSynthesize = F,
        verbosity = "INFO",
        cdmVersion = 5)
}

}
